Bleach_Prj/requirements.txt
Flask==2.2.3
google-auth==2.16.0
google-auth-oauthlib==0.5.3
google-auth-httplib2==0.1.0
mysql-connector-python==8.0.33
gunicorn==20.0.4
werkzeug==2.2.3 
APScheduler==3.8.1  
pytz==2024.1
